[OAI] Refactor script into module
