+++ src/api/oai/exceptions.py
@@ -1,0 +1,3 @@
+class OAIException(Exception):
+    pass
+
